Add an extra output that loads the grapesjs plugin on demand from a provided grapesjs instance instead of relying on a globally provided grapesjs instance.
