+++ src/index.callable.js
@@ -1,0 +1,5 @@
+import initialize from './initialize';
+
+export default function (grapesjs) {
+  initialize(grapesjs);
+}
